+++ src/apps/publisher/__main__.py
@@ -1,4 +1,4 @@
-from apps.subscriber.cli import main
+from apps.publisher.cli import main
 
 if __name__ == "__main__":
     raise SystemExit(main())