+++ src/mqtt2postgres/mqtt.py
@@ -1,4 +1,0 @@
-from broker.client import create_subscriber_client as create_mqtt_client
-from broker.client import topic_matches
-
-__all__ = ["create_mqtt_client", "topic_matches"]